--- src/screens/WalletCreatedScreen/WalletCreatedScreen.styles.ts
@@ -1,8 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-export default StyleSheet.create({
-	image: {
-		width: 276,
-		height: 336
-	}
-});
