+++ src/screens/WalletCreatedScreen/styles.ts
@@ -1,0 +1,21 @@
+import { StyleSheet } from 'react-native';
+
+export default StyleSheet.create({
+	container: {
+		flex: 1,
+		justifyContent: 'space-between'
+	},
+	image: {
+		width: 276,
+		height: 336
+	},
+	textContainer: {
+		width: '100%',
+		alignItems: 'center'
+	},
+	buttonContainer: {
+		width: '100%',
+		paddingHorizontal: 24,
+		marginBottom: 48
+	}
+});
