+++ src/screens/WalletScreen/components/Accounts/Image/index.tsx
@@ -1,30 +1,12 @@
-import { View, StyleSheet } from 'react-native';
+import { View } from 'react-native';
 import React from 'react';
 import { Icon } from '@components';
 import { useTheme } from '@hooks';
-import { IconType } from '@styles';
-
-const styles = StyleSheet.create({
-	container: {
-		justifyContent: 'center',
-		alignItems: 'center'
-	},
-	background: {
-		opacity: 0.1,
-		width: 40,
-		height: 40,
-		borderRadius: 12
-	}
-});
-
-interface ImageProps {
-	icon?: IconType;
-	alert?: boolean;
-}
+import styles from './styles';
+import { ImageProps } from './types';
 
 const Image: React.FC<ImageProps> = ({ icon = 'wallet2Stroke', alert }) => {
 	const { colors } = useTheme();
-
 	return (
 		<View style={styles.container}>
 			<View style={[styles.background, { backgroundColor: alert ? colors.alert1 : colors.alert5 }]} />
