+++ src/screens/WalletScreen/components/Accounts/Image/styles.ts
@@ -1,0 +1,14 @@
+import { StyleSheet } from 'react-native';
+
+export default StyleSheet.create({
+	container: {
+		justifyContent: 'center',
+		alignItems: 'center'
+	},
+	background: {
+		opacity: 0.1,
+		width: 40,
+		height: 40,
+		borderRadius: 12
+	}
+});
