+++ src/screens/WalletScreen/components/Accounts/Image/types.ts
@@ -1,0 +1,6 @@
+import { IconType } from '@styles';
+
+export interface ImageProps {
+	icon?: IconType;
+	alert?: boolean;
+}
