+++ src/screens/WalletScreen/components/Accounts/styles.ts
