+++ src/screens/WalletScreen/components/ActionsPanel/utils.ts
@@ -1,0 +1,11 @@
+const arr = [
+	{ name: 'Send', icon: 'sendStroke' },
+	{ name: 'Exchange', icon: 'exchangeStroke' },
+	{ name: 'Receive', icon: 'receiveStroke' },
+	{ name: 'Copy address', icon: 'copyStroke' }
+	// { name: 'New wallet', icon: 'walletStroke' },
+	// { name: 'Switch accounts', icon: 'avatarStroke' },
+	// { name: 'Delete wallet', icon: 'closeStroke' }
+];
+
+export { arr };
