+++ src/screens/WalletScreen/components/AssetsPanel/index.tsx
@@ -1,63 +1,11 @@
 import React from 'react';
-import { View, TouchableOpacity, StyleSheet, Image, GestureResponderEvent } from 'react-native';
+import { View, TouchableOpacity, Image } from 'react-native';
 import { Text, Icon } from '@components';
 import { commify } from 'ethers/lib/utils';
 import makeBlockie from 'ethereum-blockies-base64';
 import { useTheme } from '@hooks';
-
-const styles = StyleSheet.create({
-	assetsContainer: {
-		width: '100%',
-		height: 161,
-		borderRadius: 24,
-		overflow: 'hidden',
-		marginBottom: 16
-	},
-	assetsMain: {
-		flexDirection: 'row',
-		flex: 1,
-		justifyContent: 'space-between',
-		alignItems: 'center',
-		paddingHorizontal: 24,
-		paddingTop: 24,
-		paddingBottom: 17
-	},
-	buttonsContainer: {
-		flexDirection: 'row',
-		height: 53,
-		borderTopWidth: 1
-	},
-	addFundsButtonContainer: {
-		width: '50%',
-		borderRightWidth: 1,
-		flexDirection: 'row',
-		justifyContent: 'center',
-		alignItems: 'center'
-	},
-	sendButtonContainer: {
-		width: '50%',
-		borderLeftWidth: 1,
-		flexDirection: 'row',
-		justifyContent: 'center',
-		alignItems: 'center'
-	},
-
-	avatar: {
-		width: 48,
-		height: 48,
-		borderRadius: 24,
-		borderColor: 'rgba(243, 247, 255, 1)',
-		borderWidth: 4
-	}
-});
-
-interface AssetsPanelProps {
-	balance: string;
-	address: string;
-	onSave: (event: GestureResponderEvent) => void;
-	onAddFunds: (event: GestureResponderEvent) => void;
-	onWalletAssets: (event: GestureResponderEvent) => void;
-}
+import styles from './styles';
+import { AssetsPanelProps } from './types';
 
 const AssetsPanel: React.FC<AssetsPanelProps> = ({ balance, address, onAddFunds, onSave, onWalletAssets }) => {
 	const { colors } = useTheme();
