+++ src/screens/WalletScreen/components/Header/styles.ts
@@ -1,0 +1,14 @@
+import { StyleSheet } from 'react-native';
+
+export default StyleSheet.create({
+	headerContainer: {
+		flexDirection: 'row',
+		justifyContent: 'space-between',
+		alignItems: 'center',
+		marginBottom: 24
+	},
+	iconsContainer: {
+		flexDirection: 'row',
+		alignItems: 'center'
+	}
+});
