+++ src/screens/WalletScreen/components/Header/types.ts
@@ -1,0 +1,5 @@
+import { GestureResponderEvent } from 'react-native';
+
+export interface HeaderProps {
+	onSettingsPress: (event: GestureResponderEvent) => void;
+}
