+++ src/screens/WalletScreen/components/Transactions/Transaction/utils.tsx
@@ -1,0 +1,3 @@
+const truncate = (num: number | string, idx = 2) => +num.toString().slice(0, num.toString().indexOf('.') + (idx + 1));
+
+export { truncate };
