+++ src/screens/WalletScreen/components/Transactions/TransactionIcon/styles.tsx
@@ -1,0 +1,12 @@
+import { StyleSheet } from 'react-native';
+
+export const styles = StyleSheet.create({
+	circle: {
+		width: 32,
+		height: 32,
+		borderRadius: 16,
+		borderWidth: 1,
+		justifyContent: 'center',
+		alignItems: 'center'
+	}
+});
