+++ src/screens/WalletScreen/components/Transactions/TransactionIcon/types.ts
@@ -1,0 +1,5 @@
+interface TransactionIconProps {
+	received: boolean;
+}
+
+export type { TransactionIconProps };
