+++ src/screens/WalletScreen/components/Transactions/styles.tsx
@@ -1,0 +1,18 @@
+import { StyleSheet } from 'react-native';
+
+export const styles = StyleSheet.create({
+	container: {
+		flex: 1,
+		paddingHorizontal: 14,
+		paddingTop: 14
+	},
+	tableContainer: {
+		flex: 1,
+		alignItems: 'center',
+		paddingVertical: 14
+	},
+	image: {
+		width: 152,
+		height: 152
+	}
+});
