+++ src/screens/WalletScreen/components/Transactions/types.tsx
@@ -1,0 +1,7 @@
+interface TransactionsProps {
+	loading: boolean;
+	onSeeAllTransactions: () => void;
+	onAddFunds: () => void;
+}
+
+export type { TransactionsProps };
