+++ src/screens/WalletScreen/components/index.ts
@@ -1,0 +1,7 @@
+import AssetsPanel from './AssetsPanel';
+import ActionsPanel from './ActionsPanel';
+import Header from './Header';
+import Accounts from './Accounts';
+import Transactions from './Transactions';
+
+export { AssetsPanel, ActionsPanel, Header, Accounts, Transactions };
