+++ src/screens/WalletScreen/types.ts
