+++ src/screens/WelcomeScreen/ImportWalletModal/styles.tsx
@@ -1,0 +1,19 @@
+import { StyleSheet } from 'react-native';
+
+export const styles = StyleSheet.create({
+	container: {
+		paddingHorizontal: 24
+	},
+	textAreaContainer: {
+		width: '100%',
+		height: 110,
+		marginBottom: 24
+	},
+	textarea: {
+		borderRadius: 24,
+		height: 104,
+		overflow: 'hidden',
+		borderWidth: 1,
+		paddingHorizontal: 24
+	}
+});
