+++ src/screens/WelcomeScreen/ImportWalletModal/types.tsx
@@ -1,0 +1,4 @@
+export interface ImportFlowProps {
+	onImportFinished: () => void;
+	onDismiss: () => void;
+}
