+++ src/screens/WelcomeScreen/styles.ts
